requirements.txt
openai
pydantic
python-dotenv
pandas
matplotlib
tqdm
